feat: api 테스트 완료

+++ src/main/java/com/example/janghakmoney/scholarship/ScholarshipService.java
@@ -32,7 +32,7 @@ public class ScholarshipService {
         );
     }
 
-    public Scholarship findScholarshipDetail(Integer scholarshipId) {
+    public Scholarship findScholarshipDetail(Long scholarshipId) {
         return scholarshipRepository.findById(scholarshipId)
                 .orElseThrow(() -> new EntityNotFoundException("해당 장학금을 찾을 수 없습니다. ID: " + scholarshipId));
     }
